:book: [RB-162] Refactor store architecture
