--- src/client/store/about/hooks/index.js
@@ -1,1 +1,0 @@
-export { useAbout } from './useAbout';
