+++ src/client/store/about/hooks/useAbout/index.js
@@ -1,0 +1,3 @@
+import { useAbout } from './useAbout';
+
+export default useAbout;
