--- src/client/store/auth/hooks/index.js
@@ -1,1 +1,0 @@
-export { useAuth } from './useAuth';
