+++ src/client/store/auth/hooks/useAuth/index.js
@@ -1,0 +1,3 @@
+import { useAuth } from './useAuth';
+
+export default useAuth;
