--- src/client/store/config/hooks/index.js
@@ -1,1 +1,0 @@
-export { useConfig } from './useConfig';
