+++ src/client/store/config/hooks/useConfig/index.js
@@ -1,0 +1,3 @@
+import { useConfig } from './useConfig';
+
+export default useConfig;
