+++ src/client/store/contact/hooks/useContact/index.js
@@ -1,0 +1,3 @@
+import { useContact } from './useContact';
+
+export default useContact;
