--- src/client/store/csrf/hooks/index.js
@@ -1,1 +1,0 @@
-export { useCSRF } from './useCSRF';
