+++ src/client/store/csrf/hooks/useCSRF/index.js
@@ -1,0 +1,3 @@
+import { useCSRF } from './useCSRF';
+
+export default useCSRF;
