--- src/client/store/docs/hooks/index.js
@@ -1,1 +1,0 @@
-export { useDocs } from './useDocs';
