+++ src/client/store/docs/hooks/useDocs/index.js
@@ -1,0 +1,3 @@
+import { useDocs } from './useDocs';
+
+export default useDocs;
