--- src/client/store/messages/hooks/index.js
@@ -1,1 +1,0 @@
-export { useMessages } from './useMessages';
