+++ src/client/store/messages/hooks/useMessages/index.js
@@ -1,0 +1,3 @@
+import { useMessages } from './useMessages';
+
+export default useMessages;
