+++ src/client/store/messages/index.js
@@ -1,2 +1,2 @@
 export { messagesStore } from './messages.store';
-export { useMessages } from './hooks';
+export { default as useMessages } from './hooks/useMessages';
