--- src/client/store/tokens/hooks/index.js
@@ -1,1 +1,0 @@
-export { useTokens } from './useTokens';
