+++ src/client/store/tokens/hooks/useTokens/index.js
@@ -1,0 +1,3 @@
+import { useTokens } from './useTokens';
+
+export default useTokens;
