+++ src/client/store/tokens/index.js
@@ -1,2 +1,2 @@
 export { tokensStore } from './tokens.store';
-export { useTokens } from './hooks';
+export { default as useTokens } from './hooks/useTokens';
