--- src/client/store/user/hooks/index.js
@@ -1,2 +1,0 @@
-export { useUser } from './useUser';
-export { usePermissions } from './usePermissions';
