+++ src/client/store/user/hooks/usePermissions/index.js
@@ -1,0 +1,3 @@
+import { usePermissions } from './usePermissions';
+
+export default usePermissions;
