+++ src/client/store/user/hooks/useUser/index.js
@@ -1,0 +1,3 @@
+import { useUser } from './useUser';
+
+export default useUser;
